# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_transport_handler.clientLinksDTO:
    properties:
      link:
        type: string
      subscription:
        type: string
    type: object
...
  internal_transport_handler.inboundDTO:
    properties:
      clientCount:
        type: integer
      createdAt:
        type: string
      dest:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      port:
        type: integer
      protocol:
        type: string
      remark:
        type: string
      settings:
        $ref: '#/definitions/internal_transport_handler.inboundSettingsDTO'
      sni:
        type: string
      tls:
        type: string
      transmission:
        type: string
    type: object
  internal_transport_handler.inboundRequest:
    properties:
      acmeDomain:
        type: string
      acmeEmail:
        type: string
      certPath:
        type: string
      dest:
        type: string
      keyPath:
        type: string
      port:
        type: integer
      protocol:
        type: string
      remark:
        type: string
      sni:
        type: string
      tls:
        type: string
      transmission:
        type: string
    type: object
  internal_transport_handler.inboundSettingsDTO:
    properties:
      flow:
        type: string
      grpcServiceName:
        type: string
      publicKey:
        type: string
      shortId:
        type: string
      wsPath:
        type: string
    type: object
...
  internal_transport_handler.trafficPointDTO:
    properties:
      down:
        type: integer
      t:
        type: integer
      up:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Local-first web panel for managing a sing-box process.
  title: SingGrok API
  version: 0.1.0
paths:
...
  /clients/{id}/links:
    get:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.clientLinksDTO'
      security:
      - BearerAuth: []
      summary: Get a client's share link and subscription URL
      tags:
      - clients
...
  /dashboard/traffic:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_transport_handler.trafficPointDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Recent throughput history
      tags:
      - dashboard
...
  /inbounds:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_transport_handler.inboundDTO'
            type: array
      security:
      - BearerAuth: []
      summary: List inbounds
      tags:
      - inbounds
    post:
      consumes:
      - application/json
      parameters:
      - description: Inbound
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handler.inboundRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_handler.inboundDTO'
      security:
      - BearerAuth: []
      summary: Create an inbound
      tags:
      - inbounds
  /inbounds/{id}:
    delete:
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an inbound
      tags:
      - inbounds
    get:
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.inboundDTO'
      security:
      - BearerAuth: []
      summary: Get an inbound
      tags:
      - inbounds
    put:
      consumes:
      - application/json
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: integer
      - description: Inbound
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handler.inboundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.inboundDTO'
      security:
      - BearerAuth: []
      summary: Update an inbound
      tags:
      - inbounds
  /inbounds/{id}/clone:
    post:
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_handler.inboundDTO'
      security:
      - BearerAuth: []
      summary: Clone an inbound
      tags:
      - inbounds
  /inbounds/{id}/toggle:
    post:
      parameters:
      - description: Inbound ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.inboundDTO'
      security:
      - BearerAuth: []
      summary: Enable/disable an inbound
      tags:
      - inbounds
...
  /subscription/{token}:
    get:
      description: Returns the client's connection config. ?format=base64|plain|json
        (default base64).
      parameters:
      - description: Subscription token
        in: path
        name: token
        required: true
        type: string
      - description: base64 | plain | json
        in: query
        name: format
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Fetch a subscription (public)
      tags:
      - subscription
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer " followed by the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"